🔥 Remove the BufferEventTransaction in favor of OutboxEventTransaction
